# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Pottery Studio API
  version: 1.0.0
  description: API for managing pottery studio users, reservations and forum posts.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        admin:
          type: boolean

    Reservation:
      type: object
      required: [id, date, timeSlot, workplaceId]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        timeSlot:
          type: string
        workplaceId:
          type: integer
        supplies:
          type: string
        userId:
          type: integer

    Post:
      type: object
      required: [id, title, content]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        userId:
          type: integer

paths:

  /auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input
        '409':
          description: Email already exists
        '500':
          description: Server error

  /auth/login:
    post:
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '500':
          description: Server error

  /reservations:
    get:
      summary: Get reservations for a specific date
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date to filter reservations
      responses:
        '200':
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '400':
          description: Invalid date format
        '401':
          description: Unauthorized
        '500':
          description: Server error

    post:
      summary: Create reservation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [date, timeSlot, workplaceId]
              type: object
              properties:
                date:
                  type: string
                  format: date
                timeSlot:
                  type: string
                workplaceId:
                  type: integer
                supplies:
                  type: string
      responses:
        '201':
          description: Reservation created
        '400':
          description: Invalid data
        '401':
          description: Unauthorized
        '409':
          description: Time slot already booked
        '500':
          description: Server error

  /reservations/{id}:
    delete:
      summary: Cancel reservation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Reservation deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
        '500':
          description: Server error

  /posts:
    get:
      summary: Get forum posts
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Server error

    post:
      summary: Create forum post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [title, content]
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Post created
        '400':
          description: Invalid data
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /posts/{id}:
    put:
      summary: Update post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: Post updated
        '400':
          description: Invalid data
        '401':
          description: Unauthorized
        '404':
          description: Post not found
        '500':
          description: Server error

    delete:
      summary: Delete post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Post deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Post not found
        '500':
          description: Server error

  /workplaces/available:
    get:
      summary: Get available workplaces for a specific date and time
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: timeSlot
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of available workplaces
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    workplaceId:
                      type: integer
        '400':
          description: Invalid input
        '500':
          description: Server error

security:
  - BearerAuth: []